reorganized code
reorganized the code in multiple files, created the new page and the beginning of the matching system
